Fixed bugs in the MyriadEnumSet and CombinedPrFunction reader logic caused by too small line and element index counters (was I16u, should be size_t).

+++ src/cpp/core/types/MyriadEnumSet.h
@@ -167,10 +167,10 @@ public:
      * given by the \p in parameter. When reading from the \p in stream, uses
      * the \p currentLineNumber parameter to track the current line number.
      *
-     * @param currentLineNumber A reference to the current line number.
      * @param in Input stream containing the enum specification.
+     * @param currentLineNumber A reference to the current line number.
      */
-    void initialize(istream& in, UInt16& currentLineNumber);
+    void initialize(istream& in, size_t& currentLineNumber);
 
     /**
      * Return the name of this enum set.
